+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
 pub mod db;
+pub mod events;
 pub mod logic;
 pub mod utils;
